edit olympiadbench
